Render a list of airplanes on a page
Actually having the code do something (for a user on a page, as opposed
to an automated test), requires a few adjustments to the directory
structure.

Move lib into app directory, so that it is accessible from the server,
using `cd app && pythom -m SimpleHTTPServer`. The reason for using an
http server is to permit the request to api/airplanes.json - serving it
straight off the file system breaks the async request, because it lacks
certain headers.

Add requirejs to lib so we can bootstrap our application.